+++ Model/Convenience Objects/Door.cs
@@ -18,13 +18,13 @@ namespace Spectrum.Model
     {
 		public Door() : base() {
 			/* Todo: Polygon rectangle */
-			/* TODO: Door iamge */
+			/* TODO: Door image name */
 			AffectedByGravity = false;
 			Pickupable = false;
 			/* TODO: X-Event complete level */
 		}
 		/* Polygon = Rectangle, Image = DefaultDoor, Affected By Gravity = NO, Pickupable = NO, X-Event = Complete Level */
-		public Door(int id, Colors viewableColors, Vector2 position, bool affectedByGravity, Vector2 velocity, bool inactive, Texture2D inactiveImage, List<Event> events, bool existsWhenNotViewed, Level container)
+        public Door(double id, Colors viewableColors, Vector2 position, bool affectedByGravity, Vector2 velocity, bool inactive, Texture2D inactiveImage, List<Event> events, bool existsWhenNotViewed, Level container)
 			: base(id, viewableColors, null /* TODO: Rectangle polygon */, null /* TODO: Door image */, position, affectedByGravity, velocity, null, null, false, inactive, inactiveImage, events, existsWhenNotViewed, container)
 		{}
     }
